[all] More warning fix
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@5125 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux_plugins/ADM_demuxers/MpegPS/ADM_psAudioProbe.cpp
@@ -83,7 +83,7 @@ listOfPsAudioTracks *psProbeAudio(const char *fileName)
     }
 
 end:
-    printf("[PsDemux] Audio probe done, found %lu tracks\n",tracks->size());
+    printf("[PsDemux] Audio probe done, found %d tracks\n",(int)tracks->size());
     delete packet;
     
     if(tracks->size()==0)
